Elaborate on group purpose, add GH discussions link
This change adds introductional text to elaborate on the group's
raison d'être based on initial feedback we received on social media
immediately after launch.

The change also adds a reference to our newly started GH discussions
boards for engaging with the group on generic topics not related to
existing specs and docs.

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>